refactor(activity): extract AssetListControlBar + ActivityEmptyState
