+++ app/components/Views/ActivityScreen/components/ActivityEmptyState/index.ts
@@ -1,0 +1,7 @@
+export { default } from './ActivityEmptyState';
+export type { ActivityEmptyStateProps } from './ActivityEmptyState';
+export {
+  ActivityEmptyStateAction,
+  getActivityEmptyState,
+  type ActivityEmptyStateConfig,
+} from './empty-states';
