+++ app/components/Views/ActivityScreen/components/AssetListControlBar/index.ts
@@ -1,0 +1,2 @@
+export { default } from './AssetListControlBar';
+export type { AssetListControlBarProps } from './AssetListControlBar';
