+++ app/components/Views/ActivityScreen/types.ts
@@ -40,7 +40,7 @@ export type ActivityKind =
   | 'predictionCashedOut'
   | 'predictionPlaced'
   | 'perpsAddFunds'
-  | 'perpsWithdrawFunds'
+  | 'perpsWithdraw'
   | 'perpsOpenLong'
   | 'perpsCloseLong'
   | 'perpsCloseLongLiquidated'
@@ -106,7 +106,7 @@ export const ACTIVITY_TYPE_FILTER_KINDS: Record<
   ]),
   [ActivityTypeFilter.Perps]: new Set<ActivityKind>([
     'perpsAddFunds',
-    'perpsWithdrawFunds',
+    'perpsWithdraw',
     'perpsOpenLong',
     'perpsCloseLong',
     'perpsCloseLongLiquidated',
